+++ packages/ricos-plugins/web/src/resolvers/resolvers.ts
@@ -1,0 +1,5 @@
+import { TiptapContentResolver } from './ContentResolver';
+
+export const alwaysVisibleResolver = TiptapContentResolver.create('ALWAYS_VISIBLE', () => {
+  return true;
+});