+++ packages/ricos-types/web/src/contentResolverTypes.ts
@@ -1,0 +1,5 @@
+export interface IContentResolver<T> {
+  id: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  resolve: (content: T) => any;
+}